build(android): Play refused the bundle, so the Play build moves to Qt 6.10.3 with 16 KB pages and targetSdk 36
Play requires new apps to target API 36 and, above 34, to ship 16 KB aligned
native libraries. Qt 6.8.3's Android prebuilts are 4 KB aligned, so the Play
bundle is built from a separate Qt 6.10.3 + NDK r27c toolchain in
build-android-play, while the fleet APK job stays on Qt 6.6.3 for Android 6/7.

tools/android/build-play-aab.sh is the whole path: configure, AAR, bundleRelease,
then tools/android/check-16k.sh refuses any bundle with a 4 KB LOAD segment.
MapLibre moves to 11.13.5 because 11.5.2 was 4 KB. Qt 6.10 no longer exports
Qt6::*Private targets implicitly, so they are requested as optional components,
and two sources relied on a transitive include and a QFlags conversion that 6.10
dropped.

+++ src/ADSB/ADSBVehicle.cc
@@ -36,7 +36,7 @@ void ADSBVehicle::update(const ADSB::VehicleInfo_t &vehicleInfo)
         return;
     }
 
-    qCDebug(ADSBVehicleLog) << "Updating" << QStringLiteral("%1 Flags: %2").arg(vehicleInfo.icaoAddress, 0, 16).arg(vehicleInfo.availableFlags, 0, 2);
+    qCDebug(ADSBVehicleLog) << "Updating" << QStringLiteral("%1 Flags: %2").arg(vehicleInfo.icaoAddress, 0, 16).arg(vehicleInfo.availableFlags.toInt(), 0, 2);
 
     const QGeoCoordinate previous = coordinate();
     if (vehicleInfo.availableFlags & ADSB::LocationAvailable) {
